Former-commit-id: 8aac4e559190041453f6d87fe162c3e32d0660fb

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/exercise/response/ExerciseResponse.java
@@ -8,16 +8,16 @@ import lombok.*;
 @Getter
 public class ExerciseResponse {
 
-    private Long exerciseId;
+    private int exerciseId;
 
     private String name; // 운동 명
 
-    private Long setTime; // 근력 : 세트 수, 유산소 : 분
+    private int setTime; // 근력 : 세트 수, 유산소 : 분
 
-    private Long number; // 횟수
+    private int number; // 횟수
 
     private boolean success; // 운동 성공 여부
 
-    private Long weight; //중량
+    private int weight; //중량
 
 }
