+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/individualGoals/request/IndividualGoalDeleteRequest.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class IndividualGoalDeleteRequest {
 
     private  String Authorization;
-    private Long individualGoalId;
+    private int individualGoalId;
 }
