+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/repository/FeedTagRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FeedTagRepository extends JpaRepository<FeedTag,Long> {
+public interface FeedTagRepository extends JpaRepository<FeedTag,Integer> {
 
 }
